Using KDE now and will see if I like it.
Using Sway was good but I found it kind of hard to work with when it comes to
application compatibility. Some applications just do not work that well with
window managers. Never really tried KDE so will attempt to use it because I
do like to have some sort of unified environment.